+++ src/constants.ts
@@ -24,6 +24,7 @@ export const DefaultOptions: Config = {
 	popupTimer: 30,
 	skipFollowerCount: 1e6,
 	soupcanIntegration: false,
+	blockPromoted: false,
 
 	// this isn"t set, but is used
 	// TODO: when migrating to firefox manifest v3, check to see if sets can be stored yet
@@ -46,11 +47,13 @@ export const ReasonBlueVerified: number = 0;
 export const ReasonNftAvatar: number = 1;
 export const ReasonBusinessVerified: number = 2;
 export const ReasonTransphobia: number = 3;
+export const ReasonPromoted: number = 4;
 export const ReasonMap = {
 	[ReasonBlueVerified]: "Twitter Blue verified",
 	[ReasonNftAvatar]: "NFT avatar",
 	[ReasonBusinessVerified]: "Twitter Business verified",
 	[ReasonTransphobia]: "transphobia",
+	[ReasonPromoted]: "promoting tweets",
 };
 export const LegacyVerifiedUrl: string = "https://gist.githubusercontent.com/travisbrown/b50d6745298cccd6b1f4697e4ec22103/raw/012009351630dc351e3a763b49bf24fa50ca3eb7/legacy-verified.csv";
 export const Browser = chrome.runtime.getManifest()?.browser_specific_settings?.gecko === undefined ? "chrome" : "firefox";